+++ MHD/C_src/CORE_C/m_psf_data_4_viewer_c.h
@@ -65,7 +65,7 @@ struct psf_data{
 	double **color_nod;
 	
 	int nnod_added_4_map;
-	int **inod_org_4_map_itp;
+	long **inod_org_4_map_itp;
 	double **coef_4_map_itp;
     
     struct psf_edge_data_c *psf_edge;
